Remove Main_page app
