+++ Account/forms.py
@@ -1,7 +1,7 @@
 from django import forms
 from django.contrib.auth.models import User
 from django.utils import timezone
-from Account.models import Tweet
+from .models import Tweet
 
 class Generate_Tweet(forms.Form):
     tweet_text = forms.CharField(max_length=280, label=False, widget=forms.Textarea(attrs={'placeholder': "What's happening?", 'style': 'resize: vertical', 'rows': 2,}))
