+++ Account/templates/Account/homepage.html
@@ -1,6 +1,6 @@
 
 
-{% extends "Main_Page/base.html" %}
+{% extends "Account/base_homepage.html" %}
 
 <!-- Navbar -->
 {% block navbar %}
