+++ Account/templates/Account/profile.html
@@ -96,7 +96,7 @@
 {% block navbar %}
 
       <div class="topnav">
-        <a href="{% url 'Main_Page:index' %}">Home</a>
+        <a href="{% url 'Account:index' %}">Home</a>
         <a href="#">Followers {{followed_by_len}}</a>
         <a href="#">Following {{following_len}}</a>
         <a href="#">Likes</a>
