+++ Account/urls.py
@@ -4,7 +4,7 @@ from django.contrib.auth import views as auth_views
 
 app_name = 'Account'
 urlpatterns = [
-    #path('',views.index,name='index'), #TODO: Make this into a 404 or something /accounts by itself should not do anything
+    path('',views.index,name='index'), 
     path('register/',views.register,name='register'),
     path('login/',auth_views.LoginView.as_view() ,name='login'),
     path('logout/',auth_views.LogoutView.as_view(),name='logout'),
