--- Main_Page/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class MainPageConfig(AppConfig):
-    name = 'Main_Page'
