--- Main_Page/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-from . import views
-
-app_name = 'Main_Page'
-urlpatterns = [
-    path('',views.index,name='index'),
-]
-
