+++ src/urls.py
@@ -19,5 +19,5 @@ from django.urls import include, path
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('profile/',include('Account.urls')),
-    path('',include('Main_Page.urls'))
+    path('',include('Account.urls'))
 ]
